requirements.txt
tensorflow-gpu ==1.5
h5py
keras
gensim
rbfopt
